Struct scram::ClientFinal

pub struct ClientFinal {
    // some fields omitted
}

The third state of the SCRAM mechanism after the first server message was successfully processed.

Methods

impl ClientFinal

fn client_final(self) -> (ServerFinal, String)

Returns the next state and the final client message.

Call the ServerFinal::handle_server_final method to continue the SCRAM handshake.