Struct scram::ServerFirst

pub struct ServerFirst<'a> {
    // some fields omitted
}

The second state of the SCRAM mechanism after the first client message was computed.

Methods

impl<'a> ServerFirst<'a>

fn handle_server_first(self, server_first: &str) -> Result<ClientFinal, Error>

Processes the first answer from the server and returns the next state or an error. If an error is returned the SCRAM handshake is aborted.

Call the ClientFinal::client_final method to continue the handshake.

Return value

This method returns only a subset of the errors defined in Error:

Error::Protocol

Error::UnsupportedExtension